Moved the frontend into the frontend directory. Preparation for further decoupling of the frontend.
