+++ MANIFEST.in
@@ -4,3 +4,4 @@ include ez_setup.py
 include LICENSE
 include ga4gh/server/templates/*.html
 include ga4gh/server/config/oidc_auth_config.yml
+include ga4gh/server/frontend/*
