+++ ga4gh/server/frontend/flaskStatus.py
@@ -1,0 +1,146 @@
+"""
+The server status class for the Flask frontend server
+
+It should be a singleton class!
+"""
+
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import humanize
+import datetime
+import flask
+
+import ga4gh.server
+import ga4gh.schemas.protocol as protocol
+
+
+class ServerStatus(object):
+    """
+    Generates information about the status of the server for display
+    """
+    def __init__(self, flaskApp):
+        self.app = flaskApp
+        self.startupTime = datetime.datetime.now()
+
+    def getConfiguration(self):
+        """
+        Returns a list of configuration (key, value) tuples
+        that are useful for users to view on an information page.
+        Note that we should be careful here not to leak sensitive
+        information. For example, keys and paths of data files should
+        not be returned.
+        """
+        # TODO what other config keys are appropriate to export here?
+        keys = [
+            'DEBUG', 'REQUEST_VALIDATION',
+            'DEFAULT_PAGE_SIZE', 'MAX_RESPONSE_LENGTH', 'LANDING_MESSAGE_HTML'
+        ]
+
+        return [(k, self.app.config[k]) for k in keys]
+
+    def getPreciseUptime(self):
+        """
+        Returns the server precisely.
+        """
+        return self.startupTime.strftime("%H:%M:%S %d %b %Y")
+
+    def getLandingMessageHtml(self):
+        filePath = self.app.config.get('LANDING_MESSAGE_HTML')
+        try:
+            htmlFile = open(filePath, 'r')
+            html = htmlFile.read()
+            htmlFile.close()
+        except:
+            html = flask.render_template("landing_message.html")
+        return html
+
+    def getNaturalUptime(self):
+        """
+        Returns the uptime in a human-readable format.
+        """
+        return humanize.naturaltime(self.startupTime)
+
+    def getProtocolVersion(self):
+        """
+        Returns the GA4GH protocol version we support.
+        """
+        return protocol.version
+
+    def getServerVersion(self):
+        """
+        Returns the software version of this server.
+        """
+        return ga4gh.server.__version__
+
+    def getUrls(self):
+        """
+        Returns the list of (httpMethod, URL) tuples that this server
+        supports.
+        """
+        self.app.urls.sort()
+        return self.app.urls
+
+    def getDatasets(self):
+        """
+        Returns the list of datasetIds for this backend
+        """
+        return self.app.backend.getDataRepository().getDatasets()
+
+    def getVariantSets(self, datasetId):
+        """
+        Returns the list of variant sets for the dataset
+        """
+        return self.app.backend.getDataRepository().getDataset(
+            datasetId).getVariantSets()
+
+    def getFeatureSets(self, datasetId):
+        """
+        Returns the list of feature sets for the dataset
+        """
+        return self.app.backend.getDataRepository().getDataset(
+            datasetId).getFeatureSets()
+
+    def getContinuousSets(self, datasetId):
+        """
+        Returns the list of continuous sets for the dataset
+        """
+        return self.app.backend.getDataRepository().getDataset(
+            datasetId).getContinuousSets()
+
+    def getReadGroupSets(self, datasetId):
+        """
+        Returns the list of ReadGroupSets for the dataset
+        """
+        return self.app.backend.getDataRepository().getDataset(
+            datasetId).getReadGroupSets()
+
+    def getReferenceSets(self):
+        """
+        Returns the list of ReferenceSets for this server.
+        """
+        return self.app.backend.getDataRepository().getReferenceSets()
+
+    def getVariantAnnotationSets(self, datasetId):
+        """
+        Returns the list of ReferenceSets for this server.
+        """
+        # TODO this should be displayed per-variant set, not per dataset.
+        variantAnnotationSets = []
+        dataset = self.app.backend.getDataRepository().getDataset(datasetId)
+        for variantSet in dataset.getVariantSets():
+            variantAnnotationSets.extend(
+                variantSet.getVariantAnnotationSets())
+        return variantAnnotationSets
+
+    def getPhenotypeAssociationSets(self, datasetId):
+        return self.app.backend.getDataRepository().getDataset(
+            datasetId).getPhenotypeAssociationSets()
+
+    def getRnaQuantificationSets(self, datasetId):
+        """
+        Returns the list of RnaQuantificationSets for this server.
+        """
+        return self.app.backend.getDataRepository().getDataset(
+            datasetId).getRnaQuantificationSets()
