+++ ga4gh/server/frontend/frontend.py
@@ -9,7 +9,6 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import os
-import datetime
 import socket
 import urlparse
 import functools
@@ -19,7 +18,6 @@ import flask
 import flask.ext.cors as cors
 from flask import request
 from flask.ext.oidc import OpenIDConnect
-import humanize
 import werkzeug
 import oic
 import oic.oauth2
@@ -30,8 +28,8 @@ import logging
 from logging import StreamHandler
 from werkzeug.contrib.cache import FileSystemCache
 from yaml import load
+import pkg_resources
 
-import ga4gh.server
 import ga4gh.server.backend as backend
 import ga4gh.server.datamodel as datamodel
 import ga4gh.server.exceptions as exceptions
@@ -39,8 +37,7 @@ import ga4gh.server.datarepo as datarepo
 import ga4gh.server.auth as auth
 import ga4gh.server.network as network
 import ga4gh.schemas.protocol as protocol
-
-import pkg_resources
+import ga4gh.server.frontend.flaskStatus as flaskStatus
 
 MIMETYPE = "application/json"
 SEARCH_ENDPOINT_METHODS = ['POST', 'OPTIONS']
@@ -64,7 +61,7 @@ app.urls = []
 requires_auth = auth.auth_decorator(app)
 
 # Edit this for flask-oidc, the endpoints are in the client_secrets.json file
-pathLocation = '/'.join(('.', 'config', 'oidc_auth_config.yml'))
+pathLocation = '/'.join(('..', 'config', 'oidc_auth_config.yml'))
 configPath = pkg_resources.resource_filename(__name__, pathLocation)
 config = import_yaml_config(config=configPath)
 app.config.update(config["frontend"])
@@ -98,134 +95,6 @@ class NoConverter(werkzeug.routing.BaseConverter):
 
 app.url_map.converters['no'] = NoConverter
 
-
-class ServerStatus(object):
-    """
-    Generates information about the status of the server for display
-    """
-    def __init__(self):
-        self.startupTime = datetime.datetime.now()
-
-    def getConfiguration(self):
-        """
-        Returns a list of configuration (key, value) tuples
-        that are useful for users to view on an information page.
-        Note that we should be careful here not to leak sensitive
-        information. For example, keys and paths of data files should
-        not be returned.
-        """
-        # TODO what other config keys are appropriate to export here?
-        keys = [
-            'DEBUG', 'REQUEST_VALIDATION',
-            'DEFAULT_PAGE_SIZE', 'MAX_RESPONSE_LENGTH', 'LANDING_MESSAGE_HTML'
-        ]
-
-        return [(k, app.config[k]) for k in keys]
-
-    def getPreciseUptime(self):
-        """
-        Returns the server precisely.
-        """
-        return self.startupTime.strftime("%H:%M:%S %d %b %Y")
-
-    def getLandingMessageHtml(self):
-        filePath = app.config.get('LANDING_MESSAGE_HTML')
-        try:
-            htmlFile = open(filePath, 'r')
-            html = htmlFile.read()
-            htmlFile.close()
-        except:
-            html = flask.render_template("landing_message.html")
-        return html
-
-    def getNaturalUptime(self):
-        """
-        Returns the uptime in a human-readable format.
-        """
-        return humanize.naturaltime(self.startupTime)
-
-    def getProtocolVersion(self):
-        """
-        Returns the GA4GH protocol version we support.
-        """
-        return protocol.version
-
-    def getServerVersion(self):
-        """
-        Returns the software version of this server.
-        """
-        return ga4gh.server.__version__
-
-    def getUrls(self):
-        """
-        Returns the list of (httpMethod, URL) tuples that this server
-        supports.
-        """
-        app.urls.sort()
-        return app.urls
-
-    def getDatasets(self):
-        """
-        Returns the list of datasetIds for this backend
-        """
-        return app.backend.getDataRepository().getDatasets()
-
-    def getVariantSets(self, datasetId):
-        """
-        Returns the list of variant sets for the dataset
-        """
-        return app.backend.getDataRepository().getDataset(
-            datasetId).getVariantSets()
-
-    def getFeatureSets(self, datasetId):
-        """
-        Returns the list of feature sets for the dataset
-        """
-        return app.backend.getDataRepository().getDataset(
-            datasetId).getFeatureSets()
-
-    def getContinuousSets(self, datasetId):
-        """
-        Returns the list of continuous sets for the dataset
-        """
-        return app.backend.getDataRepository().getDataset(
-            datasetId).getContinuousSets()
-
-    def getReadGroupSets(self, datasetId):
-        """
-        Returns the list of ReadGroupSets for the dataset
-        """
-        return app.backend.getDataRepository().getDataset(
-            datasetId).getReadGroupSets()
-
-    def getReferenceSets(self):
-        """
-        Returns the list of ReferenceSets for this server.
-        """
-        return app.backend.getDataRepository().getReferenceSets()
-
-    def getVariantAnnotationSets(self, datasetId):
-        """
-        Returns the list of ReferenceSets for this server.
-        """
-        # TODO this should be displayed per-variant set, not per dataset.
-        variantAnnotationSets = []
-        dataset = app.backend.getDataRepository().getDataset(datasetId)
-        for variantSet in dataset.getVariantSets():
-            variantAnnotationSets.extend(
-                variantSet.getVariantAnnotationSets())
-        return variantAnnotationSets
-
-    def getPhenotypeAssociationSets(self, datasetId):
-        return app.backend.getDataRepository().getDataset(
-            datasetId).getPhenotypeAssociationSets()
-
-    def getRnaQuantificationSets(self, datasetId):
-        """
-        Returns the list of RnaQuantificationSets for this server.
-        """
-        return app.backend.getDataRepository().getDataset(
-            datasetId).getRnaQuantificationSets()
 
 
 def reset():
@@ -324,7 +193,7 @@ def configure(configFile=None, baseConfig="ProductionConfig",
         cors.CORS(app, allow_headers='Content-Type')
     except AssertionError:
         pass
-    app.serverStatus = ServerStatus()
+    app.serverStatus = flaskStatus.ServerStatus(app)
 
     app.backend = _configure_backend(app)
