--- ga4gh/server/stuff.cfg
@@ -1,1 +1,0 @@
-DEBUG=False
