+++ tests/end_to_end/test_sequence_annotations.py
@@ -8,7 +8,7 @@ from __future__ import unicode_literals
 import unittest
 import logging
 
-import ga4gh.server.frontend as frontend
+import ga4gh.server.frontend.frontend as frontend
 import tests.paths as paths
 
 import ga4gh.schemas.protocol as protocol
